+++ apps/frontend/src/components/Home/LandingBanner.tsx
