+++ apps/frontend/src/lib/components/Navbar.tsx
